feat: 상세페이지 client component 생성
